Refactor player and map classes, update dependencies, and adjust game state handling
